rename the factory
