+++ src/factories/__init__.py
@@ -1,2 +1,2 @@
-from .processor_factory import get_message_adapter
+from .message_adapter_factory import get_message_adapter
 from .service_processor_factory import get_service_processor
