+++ src/factories/message_adapter_factory.py
